refactor component structure for seo readability
